Rename more "discussion" features
They contain "discussion" in CSS selectors, so I'm not sure about them.
