+++ source/features/clean-sidebar.tsx
@@ -107,7 +107,7 @@ async function clean(): Promise<void> {
 
 void features.add({
 	id: __filebasename,
-	description: 'Hides empty sections (or just their "empty" label) in the discussion sidebar.',
+	description: 'Hides empty sections (or just their "empty" label) in the conversation sidebar.',
 	screenshot: 'https://user-images.githubusercontent.com/1402241/57199809-20691780-6fb6-11e9-9672-1ad3f9e1b827.png'
 }, {
 	include: [
