+++ source/features/sticky-conversation-sidebar.tsx
@@ -1,4 +1,4 @@
-import './sticky-discussion-sidebar.css';
+import './sticky-conversation-sidebar.css';
 import select from 'select-dom';
 import debounce from 'debounce-fn';
 import * as pageDetect from 'github-url-detection';
@@ -22,7 +22,7 @@ function deinit(): void {
 
 void features.add({
 	id: __filebasename,
-	description: 'Makes the discussion sidebar sticky.',
+	description: 'Makes the conversation sidebar sticky.',
 	screenshot: 'https://user-images.githubusercontent.com/10238474/62276723-5a2eaa80-b44d-11e9-810b-ff598d1c5c6a.gif'
 }, {
 	include: [
