+++ source/refined-github.ts
@@ -12,7 +12,7 @@ import './features/pr-approvals-count.css';
 import './features/hide-tips.css';
 import './features/hide-readme-header.css';
 import './features/hide-obvious-tooltips.css';
-import './features/clean-discussions.css';
+import './features/clean-conversations.css';
 import './features/sticky-conversation-list-toolbar.css';
 import './features/always-show-branch-delete-buttons.css';
 import './features/easier-pr-sha-copy.css';
@@ -59,10 +59,10 @@ import './features/linkify-branch-references';
 import './features/batch-open-issues';
 import './features/hide-useless-comments';
 import './features/navigate-pages-with-arrow-keys';
-import './features/discussion-links-on-repo-lists';
+import './features/conversation-links-on-repo-lists';
 import './features/global-conversation-list-filters';
 import './features/conversation-filters';
-import './features/sort-issues-by-update-time'; // Must be after global-conversation-list-filters and conversation-filters and discussion-links-on-repo-lists
+import './features/sort-issues-by-update-time'; // Must be after global-conversation-list-filters and conversation-filters and conversation-links-on-repo-lists
 import './features/pinned-issues-update-time';
 import './features/latest-tag-button';
 import './features/default-branch-button';
@@ -77,7 +77,7 @@ import './features/wait-for-build';
 import './features/hide-inactive-deployments';
 import './features/pull-request-hotkey';
 import './features/quick-review-buttons';
-import './features/format-discussion-titles';
+import './features/format-conversation-titles';
 import './features/embed-gist-inline';
 import './features/comments-time-machine-links';
 import './features/hide-issue-list-autocomplete';
@@ -122,7 +122,7 @@ import './features/hide-disabled-milestone-sorter';
 import './features/tag-changelog-link';
 import './features/link-to-file-in-file-history';
 import './features/clean-sidebar';
-import './features/sticky-discussion-sidebar';
+import './features/sticky-conversation-sidebar';
 import './features/release-download-count';
 import './features/open-issue-to-latest-comment';
 import './features/toggle-everything-with-alt';
